requirements.txt
flask
pandas
numpy
matplotlib
torch
scikit-learn
pykalman
mysql-connector-python